% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bayesGDS-deprecated}
\alias{bayesGDS-deprecated}
\alias{inv.logit}
\alias{inv.vech}
\alias{log_inv.logit}
\alias{logit}
\alias{vech}
\title{Deprecated functions}
\usage{
vech(M)

inv.vech(x)

logit(p)

inv.logit(x)

log_inv.logit(x)
}
\arguments{
  \item{M}{a matrix}

  \item{x}{A vector of conforming length}

  \item{p}{A scalar, vector or matrix, where each element
  is between 0 and 1.}

  \item{p}{A scalar, vector or matrix}

  \item{p}{A scalar, vector or matrix}
}
\value{
A vector containing the lower triangle of M, ordered
column-wise.

A k x k lower triangular matrix

result = log(p/(1-p))

result = exp(x)/(1+exp(x))

result = log[exp(x)/(1+exp(x))]
}
\description{
These functions were in earlier versions, but will no
longer be maintained in this package.  They will likely be
moved to another package a some time.

vech operator on a square matrix

inverse vech operator on a vector

Logit transformation

Inverse logit transformation

Log inverse logit transformation
}
\details{
Returns a lower triangular matrix, with elements determined
by x. x must be a vector of length k(k+1)/2, where k is the
number of rows (and columns) of the result.
}

