\name{bayesGDS-package}
\alias{bayesGDS-package}
\alias{bayesGDS}
\docType{package}
\title{
Functions to support implementation of the Braun and Damien (2011) Generalized Direct Sampling algorithm.
}
\description{
These functions will be useful in running the Braun and Damien (2011) Generalized Direct Sampling algorithm.  See the vignette for a guide on how to fit these functions together.
}
\details{
\tabular{ll}{
Package: \tab bayesGDS\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2011-09-27\cr
License: \tab see LICENSE file\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Michael Braun

Maintainer: Michael Braun <braunm@mit.edu>
}
\references{
Braun, Michael and Paul Damien (2011) Generalizaed Direct Sampling for Bayesian Hierarchical Models.  http://arxiv.org/abs/1108.2245v2
}

