% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binary-data.R
\docType{data}
\name{binary-data}
\alias{binary-data}
\alias{binary_large}
\alias{binary_large-data}
\alias{binary_small}
\alias{binary_small-data}
\title{Sample simulated data for binary choice model in vignette}
\description{
Simulated data, with $N=20$ and $k=2$.  See vignette.
Generated from data-raw/binary.R

Simulated data, with $N=800$ and $k=3$.  See vignette.
Generated from data-raw/binary.R
}

