\name{get.cutoffs}
\alias{get.cutoffs}
\title{Draw thresholds for the accept-reject stage of the GDS sampling algorithm.}
\description{
Returns a vector of log(u), where u is the threshold to determine if a proposal draw should be accepted as a draw from the target posterior distribution.
	}
\usage{
get.cutoffs(log.phi, n.draws)
}

\arguments{
	\item{log.phi}{Vector of log.phi from the proposal draws.  All must be non-positive.}
	\item{n.draws}{an integer.  number of draws to be taken from the target posterior.}
}
\value{
a numeric vector for v = -log.u (the thresholds for the accept-reject stage).
}
\details{
For use in conjunction with the Braun and Damien (2012) Generalized Direct Sampling algorithm.  This is usually not called directly, since it is called from the sample.GDS function.
}






 


