% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bayesGDS-package.R
\docType{package}
\encoding{UTF-8}
\name{bayesGDS-package}
\alias{bayesGDS}
\alias{bayesGDS-package}
\title{Braun and Damien Algorithm for Scalable Rejection Sampling}
\description{
Braun and Damien Algorithm for Scalable Rejection Sampling
}
\details{
This package provides helper functions for implementing the
Braun and Damien Scalable Rejection Sampling algorthm (see
references).
}
\keyword{package}

