% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesGDS-package.R
\docType{package}
\encoding{UTF-8}
\name{bayesGDS-package}
\alias{bayesGDS}
\alias{bayesGDS-package}
\title{Braun and Damien Algorithm for Scalable Rejection Sampling}
\description{
Functions for implementing
the Braun and Damien (BD) Scalable Rejection Sampling algorthm
}
\references{
Braun, Michael and Paul Damien (2015).  Scalable Rejection Sampling for Bayesian Hierarchical Models. Marketing Science. Articles in Advance.  http://doi.org/10.1287/mksc.2014.0901
}
\keyword{package}

