\name{bayesGDS-package}
\alias{bayesGDS-package}
\alias{bayesGDS}
\docType{package}
\title{
Functions to support implementation of the Braun and Damien (2013) Generalized Direct Sampling algorithm.
}
\description{
These functions will be useful in running the Braun and Damien (2013) Generalized Direct Sampling algorithm.  The latest version of this paper is available in the package documentation.
}
\details{
\tabular{ll}{
Package: \tab bayesGDS\cr
Type: \tab Package\cr
Version: \tab 0.6-0\cr
Date: \tab 2013-12-13\cr
License: \tab see LICENSE file\cr
LazyLoad: \tab yes\cr
}

}
\author{
Michael Braun

Maintainer: Michael Braun <braunm@smu.edu>
}
